入参由表单形式改成json

+++ src/main/java/com/joyzone/platform/core/service/DynamicSerivce.java
@@ -65,7 +65,7 @@ public class DynamicSerivce extends BaseService<DynamicModel> {
         return dynamicMapper.getUserDynamicList(userId);
     }
 
-    public List<IndexDynamicListDto> getIndexDynamicList(Long userId, Long type){
+    public List<IndexDynamicListDto> getIndexDynamicList(Long userId, Integer type){
         return dynamicMapper.getIndexDynamicList();
     }
